# backend/docs/swagger/complete-swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Vertex Markets API",
    "version": "1.0.0",
    "description": "Production-ready backend API for a modern Forex/Trading company website. Provides authentication, user management, CMS content management, markets, trading platforms, blog, contact management, newsletter subscriptions, careers/jobs, partner management, file uploads, admin dashboard, and global search capabilities.",
    "contact": {
      "name": "Vertex Markets Development Team"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    },
    {
      "url": "https://api.vertexmarkets.com",
      "description": "Production server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token obtained from POST /api/v1/login"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "accessToken",
        "description": "JWT access token stored in HTTP-only cookie"
      }
    },
    "schemas": {
...
      "ContactCreateRequest": {
        "type": "object",
        "required": ["fullName", "email", "phone", "country", "subject", "message"],
        "properties": {
          "fullName": { "type": "string", "minLength": 2, "maxLength": 200, "example": "John Doe" },
          "email": { "type": "string", "format": "email", "example": "john@example.com" },
          "phone": { "type": "string", "minLength": 7, "maxLength": 25, "example": "+1234567890" },
          "country": { "type": "string", "minLength": 2, "maxLength": 100, "example": "United States" },
          "subject": { "type": "string", "minLength": 3, "maxLength": 150, "example": "Account Inquiry" },
          "message": { "type": "string", "minLength": 10, "maxLength": 4000, "example": "I would like to know more about your services." },
          "department": { "type": "string", "maxLength": 100, "example": "Support" }
        }
      },
      "ContactReplyRequest": {
        "type": "object",
        "required": ["replyMessage"],
        "properties": {
          "replyMessage": { "type": "string", "minLength": 1, "maxLength": 4000, "example": "Thank you for your inquiry..." },
          "fromEmail": { "type": "string", "format": "email", "example": "support@vertexmarkets.com" },
          "assignedTo": { "type": "string", "example": "user_xyz789" }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/v1/contact": {
      "post": {
        "tags": ["Contact"],
        "summary": "Submit a contact message",
        "description": "Submit a new contact inquiry. Public endpoint. All required fields must be provided.",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ContactCreateRequest" }
            }
          }
        },
        "responses": {
          "201": { "description": "Contact message created" },
          "400": { "description": "Validation failed" }
        }
      },
      "get": {
        "tags": ["Contact"],
        "summary": "List contact messages (admin)",
        "description": "Retrieve a paginated list of all contact messages. Requires Admin role with contact:write permission.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "query", "name": "page", "schema": { "type": "integer", "minimum": 1 } },
          { "in": "query", "name": "limit", "schema": { "type": "integer", "minimum": 1, "maximum": 100 } },
          { "in": "query", "name": "sortBy", "schema": { "type": "string", "enum": ["createdAt", "updatedAt", "status", "repliedAt"] } },
          { "in": "query", "name": "sortOrder", "schema": { "type": "string", "enum": ["asc", "desc"] } },
          { "in": "query", "name": "q", "schema": { "type": "string" } },
          { "in": "query", "name": "status", "schema": { "type": "string", "enum": ["new", "in_progress", "replied", "closed"] } },
          { "in": "query", "name": "assignedTo", "schema": { "type": "string" } },
          { "in": "query", "name": "active", "schema": { "type": "boolean" } }
        ],
        "responses": {
          "200": { "description": "List of contacts" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" }
        }
      }
    },
    "/api/v1/contact/status/{id}": {
      "get": {
        "tags": ["Contact"],
        "summary": "Get contact status (public)",
        "description": "Public endpoint to check the status of a submitted contact message by its ID.",
        "security": [],
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "string" }, "description": "Contact ID" }
        ],
        "responses": {
          "200": { "description": "Contact status" },
          "404": { "description": "Contact not found" }
        }
      }
    },
    "/api/v1/contact/{id}": {
      "get": {
        "tags": ["Contact"],
        "summary": "Get contact by ID (admin)",
        "description": "Retrieve a specific contact message by its ID.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Contact details" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" },
          "404": { "description": "Contact not found" }
        }
      },
      "put": {
        "tags": ["Contact"],
        "summary": "Replace contact (admin)",
        "description": "Fully replace a contact message.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ContactCreateRequest" }
            }
          }
        },
        "responses": {
          "200": { "description": "Contact updated" },
          "400": { "description": "Validation failed" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" },
          "404": { "description": "Contact not found" }
        }
      },
      "patch": {
        "tags": ["Contact"],
        "summary": "Partially update contact (admin)",
        "description": "Partially update a contact message.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "fullName": { "type": "string" },
                  "email": { "type": "string", "format": "email" },
                  "phone": { "type": "string" },
                  "country": { "type": "string" },
                  "subject": { "type": "string" },
                  "message": { "type": "string" },
                  "department": { "type": "string" },
                  "active": { "type": "boolean" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Contact patched" },
          "400": { "description": "Validation failed" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" },
          "404": { "description": "Contact not found" }
        }
      },
      "delete": {
        "tags": ["Contact"],
        "summary": "Soft-delete contact (admin)",
        "description": "Soft-delete a contact message, marking it as inactive.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Contact deleted" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" },
          "404": { "description": "Contact not found" }
        }
      }
    },
    "/api/v1/contact/{id}/status": {
      "patch": {
        "tags": ["Contact"],
        "summary": "Update contact status (admin)",
        "description": "Update the status of a contact message (new, in_progress, replied, closed).",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["status"],
                "properties": {
                  "status": { "type": "string", "enum": ["new", "in_progress", "replied", "closed"] }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Status updated" },
          "400": { "description": "Validation failed" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" },
          "404": { "description": "Contact not found" }
        }
      }
    },
    "/api/v1/contact/{id}/assign": {
      "patch": {
        "tags": ["Contact"],
        "summary": "Assign contact (admin)",
        "description": "Assign a contact message to a team member for handling.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "assignedTo": { "type": "string", "description": "User ID to assign to" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Contact assigned" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" },
          "404": { "description": "Contact not found" }
        }
      }
    },
    "/api/v1/contact/{id}/reply": {
      "post": {
        "tags": ["Contact"],
        "summary": "Reply to contact (admin)",
        "description": "Send a reply to a contact inquiry. Updates the status to 'replied'.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ContactReplyRequest" }
            }
          }
        },
        "responses": {
          "200": { "description": "Reply sent" },
          "400": { "description": "Validation failed" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" },
          "404": { "description": "Contact not found" }
        }
      }
    },
...
  },
...
}